feat: add coupon UI
